Consolidating all Add pages.
