+++ src/components/addPages/ProfessorAddPage.js
@@ -12,11 +12,11 @@ import { Box, InputLabel, FormControl, MenuItem, Select, Chip, OutlinedInput, Te
 import { React, useState } from 'react';
 import {FaTimes} from 'react-icons/fa';
 
-import './../assets/styles/HomePage.css';
-import './../assets/styles/SideNav.css';
-import './../assets/styles/AddPages.css';
-import SideNavigation from './SideNavigation.js';
-import TopBar from './TopBar.js'
+import './../../assets/styles/HomePage.css';
+import './../../assets/styles/SideNav.css';
+import './../../assets/styles/AddPages.css';
+import SideNavigation from './../SideNavigation.js';
+import TopBar from './../TopBar.js'
 
 
 // Styling
